% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_diagnostics.R
\name{print.reprostat}
\alias{print.reprostat}
\title{Print a reprostat object}
\usage{
\method{print}{reprostat}(x, ...)
}
\arguments{
\item{x}{A \code{reprostat} object.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a reprostat object
}
\examples{
set.seed(1)
d <- run_diagnostics(mpg ~ wt + hp, data = mtcars, B = 20)
print(d)

}
